O dashboard vira servidor, o Tauri sai inteiro, e as dependências ficam em dia

+++ apps/cli/build.rs
@@ -104,7 +104,9 @@ fn build_date() -> String {
 fn civil_from_days(z: i64) -> (i64, u32, u32) {
     let z = z + 719_468;
     let era = if z >= 0 { z } else { z - 146_096 } / 146_097;
-    let doe = (z - era * 146_097) as i64; // [0, 146096]
+    // No cast: `z` and `era` are already `i64`. The `as` was a leftover from
+    // Hinnant's C++ original, where this line changes type to `unsigned`.
+    let doe = z - era * 146_097; // [0, 146096]
     let yoe = (doe - doe / 1460 + doe / 36_524 - doe / 146_096) / 365; // [0, 399]
     let y = yoe + era * 400;
     let doy = doe - (365 * yoe + yoe / 4 - yoe / 100); // [0, 365]
